+++ packages/@pufflig/ps-nodes-config/src/core/index.ts
@@ -1,7 +1,15 @@
-import { inputNodeConfig } from "./input";
-import { outputNodeConfig } from "./output";
+import { forinNodeConfig, forinNodeType } from "./forin";
+import { inputNodeConfig, inputNodeType } from "./input";
+import { outputNodeConfig, outputNodeType } from "./output";
 
 export const coreNodes = {
-  "core/input": inputNodeConfig,
-  "core/output": outputNodeConfig,
+  [inputNodeType]: inputNodeConfig,
+  [outputNodeType]: outputNodeConfig,
+  [forinNodeType]: forinNodeConfig,
+};
+
+export const coreNodeTypes = {
+  forinNodeType,
+  inputNodeType,
+  outputNodeType,
 };