+++ packages/@pufflig/ps-nodes-config/src/core/output.ts
@@ -1,5 +1,7 @@
 import { NodeConfig } from "@pufflig/ps-types";
 
+export const outputNodeType = "core/output" as const;
+
 export const outputNodeConfig: NodeConfig = {
   name: "End",
   description: "Final node in a workflow",